refactor analysis classes to use AnalysisCTI
